(UPDATE) Demo styling
