+++ demos/app/navbar/sample.component.css
@@ -1,0 +1,14 @@
+.sample-content {
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    grid-template-rows: auto-fill;
+}
+
+.navbar-sample {
+    height: 360px;
+    box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.2),
+0px 1px 1px 0px rgba(0, 0, 0, 0.14),
+0px 2px 1px -1px rgba(0, 0, 0, 0.12);
+    border-radius: 2px;
+    background: url('demos/app/rsrcs/sample-background.svg') no-repeat;
+    background-size: 100%;
+}
