--- demos/app/navbar/sample.html
@@ -1,5 +1,0 @@
-<igx-navbar [title]="currentView"
-            actionButtonIcon="arrow_back"
-            [isActionButtonVisible]="canGoBack()"
-            (onAction)="navigateBack()">
-</igx-navbar>
