+++ demos/app/navbar/sample.module.ts
@@ -1,14 +1,16 @@
 import { NgModule } from "@angular/core";
 
-import { IgxNavbarModule } from "../../../src/main";
+import { IgxNavbarModule } from "../../../src/main"
+import { IgxIconModule } from "../../../src/main";
 import { NavbarSampleComponent } from './sample.component';
 
 @NgModule({
     imports: [
         IgxNavbarModule,
+        IgxIconModule
     ],
     declarations: [
-        NavbarSampleComponent,
+        NavbarSampleComponent
     ]
 })
 export class NavbarSampleModule {}
