+++ demos/app/range/sample.component.css
@@ -1,0 +1,3 @@
+.sample-content {
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+}
