+++ demos/app/snackbar/sample.component.css
@@ -1,32 +1,35 @@
+.sample-content {
+    grid-template-columns: 1fr;
+    grid-template-rows: 480px;
+}
+
+.snackbar-sample {
+    position: relative;
+    height: 100%;
+    border-radius: 2px;
+    box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.2),
+0px 1px 1px 0px rgba(0, 0, 0, 0.14),
+0px 2px 1px -1px rgba(0, 0, 0, 0.12);
+    overflow: hidden;
+}
+
 .color-holder {
-    margin: 16px;
-    height: 20%;
     display: flex;
     justify-content: center;
     align-items: center;
-    text-align: center;
+    height: inherit;
     color: rgba(255,255,255, 90);
+    font-size: 1.5em;
+    padding: 16px;
     text-shadow: 0px 2px rgba(0, 0, 0, 0.2);
-    box-shadow: 0px 2px 4px -1px rgba(0, 0, 0, 0.2),
-            0px 4px 5px 0px rgba(0, 0, 0, 0.14),
-            0px 1px 10px 0px rgba(0, 0, 0, 0.12);
-    transition: background .35s cubic-bezier(0.0, 0.0, 0.2, 1);
-    
-}
-[igxButton] {
-    filter: none;
+    transition: all .35s cubic-bezier(0.0, 0.0, 0.2, 1);
+    cursor: pointer;
+    user-select: none;
+
 }
 
-h2 {
-    padding: 16px;
-}
-.color-btn {
-    position: relative;
-    margin: 16px;
-    margin-left: calc(100% - 72px);
-}
 .snackbar-container {
+    position: absolute;
     width: 100%;
-    position: fixed;
     bottom: 0;
 }
