+++ demos/app/snackbar/sample.component.html
@@ -1,24 +1,31 @@
-<div class="color-holder" [style.backgroundColor]="color">
-    <h2 igxLabel>Click the button bellow to change my color.</h2>
-</div>
-
-
-<div class="snackbar-container">
-    
-    <div class="color-btn">
-        <button 
-            igxButton="fab"
-            igxButtonBackground="mediumpurple"
-            igxButtonColor="white"
-            (click)="changeColor(snackbar)">
-            <i class="material-icons">format_paint</i>
-        </button>
+<div class="sample-wrapper">
+    <div class="sample-header">
+        <h1 class="title">Snackbar</h1>
+        <p class="description">
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas pretium lorem id tristique sagittis. Proin in mauris ac
+            tortor placerat sollicitudin eu sit amet leo. Aliquam consectetur eu sem vitae egestas. Class aptent taciti sociosqu
+            ad litora torquent per conubia nostra, per inceptos himenaeos. Cras a erat sagittis, iaculis urna ac, mollis
+            elit. Vestibulum accumsan ex at mauris sagittis, sed tincidunt justo fringilla. Phasellus vel porttitor odio.
+            Aenean quis nisi non ex ultrices efficitur. Interdum et malesuada fames ac ante ipsum primis in faucibus. Donec
+            nec felis cursus erat mattis venenatis. Curabitur congue ullamcorper nisi nec blandit. Fusce molestie consectetur
+            urna sed feugiat.
+        </p>
     </div>
 
-    <igx-snackbar #snackbar 
-                (animationStarted)="onAnimation($event)"
-                (animationDone)="onAnimation($event)"
-                [message]="message"
-                [actionText]="actionName"
-                (onAction)="undoColorChange(snackbar)"></igx-snackbar>
+    <section class="sample-content">
+        <article>
+            <h4 class="sample-title">Snackbar w/ action button</h4>
+            <div class="snackbar-sample">
+                <div class="color-holder" [style.backgroundColor]="color" (click)="changeColor(snackbar)">
+                    <span>Click me to change my color.</span>
+                </div>
+
+                <div class="snackbar-container">
+                    <igx-snackbar #snackbar [actionText]="actionName" [message]="message" (animationStarted)="onAnimation($event)" (animationDone)="onAnimation($event)"
+                        (onAction)="undoColorChange(snackbar)">
+                    </igx-snackbar>
+                </div>
+            </div>
+        </article>
+    </section>
 </div>
