+++ src/themes/material/extends/_avatar.scss
@@ -3,6 +3,7 @@
     justify-content: center;
     align-items: center;
     position: relative;
+    pointer-events: none;
 }
 
 %igx-avatar--rounded {
