+++ src/themes/material/extends/_progress.scss
@@ -7,6 +7,7 @@
     flex-flow: column nowrap;
     width: 100%;
     height: 5px;
+    margin-bottom: 32px;
     background: rgba($primary-color, .1);
 }
